# backend/openapi.yaml
openapi: 3.0.0
info:
  title: GamePlatform API
  version: v1.0.0
  description: API for managing a gaming platform including championships, events, user registrations, results, and more.
servers:
  - url: /api # Assuming all backend routes are prefixed with /api
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: A human-readable error message.
        details:
          type: string
          description: Optional additional details about the error.
        code:
          type: string
          description: Optional error code.
...
    Championship:
      type: object
      properties:
        id:
          type: string # Changed to string to match backend which uses default UUID from Supabase
          format: uuid
          readOnly: true
        name:
          type: string
        description:
          type: string
          nullable: true
        start_date:
          type: string
          format: date-time
          nullable: true
        end_date:
          type: string
          format: date-time
          nullable: true
        rules:
          type: string
          nullable: true
        prize_pool:
          type: string
          nullable: true
        status:
          type: string
          enum: [upcoming, ongoing, completed, cancelled]
          default: upcoming
        created_by:
          type: string
          format: uuid
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        # game_id: # Assuming game_id is part of your model
        #   type: integer 
        #   nullable: true
      required:
        - name
    ChampionshipCreation:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        start_date:
          type: string
          format: date-time
          nullable: true
        end_date:
          type: string
          format: date-time
          nullable: true
        # game_id:
        #   type: integer
        #   nullable: true
        rules:
          type: string
          nullable: true
        prize_pool:
          type: string
          nullable: true
        status:
          type: string
          enum: [upcoming, ongoing, completed, cancelled]
          default: upcoming
      required:
        - name

    # --- Event Schemas (Example - User can expand) ---
...
    Comment:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        user_id:
          type: string
          format: uuid
          readOnly: true # Set by backend based on auth
        content:
          type: string
        entity_type:
          type: string
        entity_id_int:
          type: integer
          nullable: true
        entity_id_uuid:
          type: string
          format: uuid
          nullable: true
        parent_id:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        user: # Embedded user info
          $ref: '#/components/schemas/UserProfileShort' # Define UserProfileShort if needed
        reply_count:
            type: integer
            readOnly: true
      required:
        - content
        - entity_type
        # entity_id_int or entity_id_uuid required based on entity_type
    CommentCreation:
      type: object
      properties:
        content:
          type: string
        entity_type:
          type: string
          description: e.g., 'championship', 'event', 'user_profile'
        entity_id_int:
          type: integer
          nullable: true
          description: Required if entity_type refers to an entity with an integer ID.
        entity_id_uuid:
          type: string
          format: uuid
          nullable: true
          description: Required if entity_type refers to an entity with a UUID ID.
        parent_id:
          type: integer
          nullable: true
          description: ID of the parent comment if this is a reply.
      required:
        - content
        - entity_type
        # Plus one of entity_id_int or entity_id_uuid, enforced by backend logic

    # Add schemas for Track, Vehicle, Rule, Result as needed, similar to Championship/Event
    # Example for a short user profile for embedding in other responses
    UserProfileShort:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          nullable: true
        avatar_url:
          type: string
          format: url
          nullable: true
  securitySchemes:
    bearerAuth: # Can be named anything
      type: http
      scheme: bearer
      bearerFormat: JWT # Supabase uses JWT
security:
  - bearerAuth: [] # Applies Bearer token auth globally to all operations by default. Override locally if needed.
paths:
...
  /championships:
    get:
      summary: Get all championships
      tags: [Championships]
      # No security for public listing, or make it optional if some data is restricted
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: A list of championships.
          content:
            application/json:
              schema:
                type: object
                properties:
                  championships:
                    type: array
                    items:
                      $ref: '#/components/schemas/Championship'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new championship (Admin only)
      tags: [Championships, Admin]
      security:
        - bearerAuth: [] # Requires admin role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChampionshipCreation'
      responses:
        '201':
          description: Championship created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Championship'
        '400':
          description: Invalid input data.
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden (user is not an admin).
  /championships/{id}:
    get:
      summary: Get a single championship by ID
      tags: [Championships]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string # Changed to string for UUID
            format: uuid
      responses:
        '200':
          description: Details of the championship.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Championship'
        '404':
          description: Championship not found.
    put:
      summary: Update an existing championship (Admin only)
      tags: [Championships, Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string # Changed to string for UUID
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChampionshipCreation' # Can be same as creation or a specific Update schema
      responses:
        '200':
          description: Championship updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Championship'
        '400':
          description: Invalid input.
        '404':
          description: Championship not found.
    delete:
      summary: Delete a championship (Admin only)
      tags: [Championships, Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string # Changed to string for UUID
            format: uuid
      responses:
        '204':
          description: Championship deleted successfully.
        '404':
          description: Championship not found.
          
  # --- Event Routes (Example - User can expand) ---
...
  /comments:
    post:
      summary: Create a new comment
      tags: [Comments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreation'
      responses:
        '201':
          description: Comment created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
    get:
      summary: Get comments for an entity
      tags: [Comments]
      parameters:
        - name: entity_type
          in: query
          required: true
          schema: { type: string }
        - name: entity_id_int
          in: query
          required: false # Required based on entity_type
          schema: { type: integer }
        - name: entity_id_uuid
          in: query
          required: false # Required based on entity_type
          schema: { type: string, format: uuid }
        - name: parent_id
          in: query
          required: false
          schema: { type: integer, nullable: true }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: A list of comments.
          content:
            application/json:
              schema:
                type: object
                properties:
                  comments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Comment'
                  total: { type: integer }
                  # ... pagination fields
  # ... other comment routes (/comments/{id} PUT, DELETE)

  # TODO: Add paths for Tracks, Vehicles, Rules, Results similar to Championships/Events
  # For example:
  # /tracks: GET, POST (Admin)
  # /tracks/{id}: GET, PUT (Admin), DELETE (Admin)
  # /vehicles: GET, POST (Admin)
  # /vehicles/{id}: GET, PUT (Admin), DELETE (Admin)
  # /rules: GET, POST (Admin)
  # /rules/{id}: GET, PUT (Admin), DELETE (Admin)
  # /results: GET, POST (Admin)
  # /results/{id}: GET, PUT (Admin), DELETE (Admin)
  
  # Example for a test protected route (if you keep it)
...